features added

+++ app/src/db/models/Statistics.ts
@@ -37,63 +37,63 @@ export class Statistics {
   })
   ThreeLetterSymbol: string;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   Infection_Risk: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   Case_Fatality_Rate: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   Test_Percentage: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   Recovery_Proporation: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   TotalCases: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   NewCases: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   TotalDeaths: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   NewDeaths: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int8', default: 0 })
   TotalRecovered: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   NewRecovered: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   ActiveCases: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int8', default: 0 })
   TotalTests: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int8', default: 0 })
   Population: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   one_Caseevery_X_ppl: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   one_Deathevery_X_ppl: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   one_Testevery_X_ppl: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   Deaths_1M_pop: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'int', default: 0 })
   Serious_Critical: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   Tests_1M_Pop: number;
 
-  @Column({ nullable: true, type: 'float8', default: 0 })
+  @Column({ nullable: true, type: 'float', default: 0 })
   TotCases_1M_Pop: number;
 }
